requirements.txt
nextcord==2.5.0
pyyaml==6.0.1